smart_room-(IoT_Platform)/requirements.txt
anytree==2.8.0
apparmor==3.0.3
appdirs==1.4.4
application-utility==1.3.2
argon2-cffi==21.1.0
asn1crypto==1.4.0
async-generator==1.10
attrs==21.3.0
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
bleach==4.1.0
btrfsutil==5.15.1
CacheControl==0.12.6
cached-property==1.5.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
Charon==1.0
click==8.0.3
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==36.0.1
cupshelpers==1.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
distro==1.6.0
dnspython==2.1.0
docker==5.0.3
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
entrypoints==0.3
fett==0.3.2
Flask==2.0.2
gps==3.23.1
html5lib==1.1
httplib2==0.20.3
idna==3.3
ifaddr==0.1.7
importlib-metadata==4.8.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isc==2.0
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.4.0
keyutils==0.6
lensfun==0.3.95
LibAppArmor==3.0.3
libtorrent===1.2.15-build-libtorrent-rasterbar-src-libtorrent-rasterbar-1.2.15-bindings-python
lit==13.0.0.dev0
louis==3.20.0
lxml==4.6.5
manjaro-sdk==0.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
meson==0.60.3
mistune==0.8.4
more-itertools==8.10.0
msgpack==1.0.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
netsnmp-python==1.0a1
networkx==2.6.3
notebook==6.4.4
npyscreen==4.10.5
numpy==1.21.5
openshot-qt==2.6.1
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.23.2
paho-mqtt==1.6.1
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
ply==3.11
progress==1.6
prometheus-client==0.11.0
prompt-toolkit==3.0.24
psutil==5.8.0
ptyprocess==0.7.0
pwquality==1.4.4
pyalsa==1.2.6
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycups==2.0.1
pycurl==7.44.1
Pygments==2.10.0
PyGObject==3.42.0
pymongo==3.12.3
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.0
pyrsistent==0.18.0
pyserial==3.5
PySide6==6.2.2
pysmbc==1.0.23
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jsonrpc-server==0.3.4
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.2.1
rados==2.0.0
rbd==2.0.0
reportlab==3.6.3
requests==2.26.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
rrdtool==0.1.10
scipy==1.7.3
scour==0.38.2
SecretStorage==3.3.1
Send2Trash==1.8.0
Shapely==1.7.1
shiboken6==6.2.2
sip==4.19.25
six==1.16.0
smbus==1.1
snooty-lextudio==1.11.6
TBB==0.2
team==1.0
terminado==0.11.1
testpath==0.5.0
texttable==1.6.4
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.0
trimesh==3.9.31
typing-extensions==4.0.1
udiskie==2.4.0
ujson==1.35
urllib3==1.26.7
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.0.2
widgetsnbextension==3.5.1
zeroconf==0.36.13
zipp==3.6.0
